Update for 1.12.13.
It was much simpler to update to this newer version of cvs than to fix the tag/tagexpand for $MidnightBSD$ type tags.

This is based on the patches from DragonFly with most of their extensinos.  The manpages aren't updated yet.

One caveat, the __fpending libc/stdio function is missing in MidnightBSD.  This is a Solarisism which was added to Linux and DragonFly.  For now, we've simply deleted references to it.  That means the pending buffer is not checked which may or may not be a real issue.

src/contrib/cvs/lib/getdate.y will not build with our system yacc.  It requires a version of bison newer than 1.75 (1.875 worked).  I will just commit the generated files instead.