Let the fun begin.
Rebranded man page (mostly).  There are still several errors and references to things like pesky alpha support.

Created a temporary hack to (hopefully) get BSD cpio working with the installation media.  If this works, we'll need to make some adjustments as to not waste memory and/or replace the mediaBufferSize function with something that makes sense.