Duh.  Switch back to /usr/local as cpp will conflict with gcc.
We might need to hack this and/or make a wrapper for the system compiler if we want to connect this to the build.

This is a rare case where we're writing to /usr/local and it's not a port.  (maybe the only time)  i don't like this one bit.