Rewrite the upgrade logic to try to upgrade packages from a bottom up order using recursion. This is necessary because package install scripts may set things up and assume that their downdepends have run.
This is one of two problems discovered during the "stargazer" test. The second is related to ports with the same name but different versions.  I've modified the ports system to pay attention to the pkgnamesuffix which will help in some cases, but say you have
python-2.4 and python-2.5.  These are distinct ports and can be installed at the same time. One idea is to use the CONFLICTS secions to know if it's safe to install two ports at the same time.  if it is, assume 2.4 doesn't get updated to 2.5.  Their presense in the index is also a clue, but the version check logic isn't that smart yet.  This will be dealt with in a later update.

I've left in some print statements as this code is untested until a new magus run can be done on i386. Use at your own risk!