Introduce CPU Affinity in MidnightBSD.
The ULE scheduler is now default on i386 and amd64 architectures.

cpuset(1) allows the user to limit a process to a group of CPU cores to run on using new system calls.

32bit versions of the system calls were also added to support 32bit binaries running on amd64 boxes. libc's
symbol map was modified to include the cpuset* system calls to expose them to userland.

This is based on work by Jeff Roberson from FreeBSD 7.1.