The atkbd(4) driver, together with the atkbdc(4) driver, provides access to the AT 84 keyboard or the AT enhanced keyboard which is connected to the AT keyboard controller.  The driver is required for the console driver syscons(4) or vt(4).  The driver exposes its own ioctl(2) interface to allow it to be configured from userland through the kbdcontrol(1) utility.
Incorrect signedness comparison in the ioctl(2) handler allows a malicious
local user to overwrite a portion of the kernel memory.

A local user may crash the kernel, read a portion of kernel memory and
execute arbitrary code in kernel context.  The result of executing an
arbitrary kernel code is privilege escalation.

Obtained from: FreeBSD