cron(8): convert vfork() usage to fork()
vfork() is error-prone, and the usage here definitely grew to not be
clearly OK given vfork-semantics; e.g. setusercontext(3) within the child.

Rip out vfork() and the rest of the references to it. fork is heavier, but
it's unclear that the difference will be all that obvious.

cron(8): rip out do_univ

This was an old Dynix hack, the function is a NOP on BSD. We have no
need to retain this; Dynix was discontinued long ago.