Callouts may be bound to a specific CPU, in which case that CPU is responsible for raising the timer interrupt which schedules execution of the callout.
A kernel thread may attempt to stop a callout while it is actively executing,
in which case the thread goes to sleep until execution has completed.  In the
meantime the callout may be re-scheduled and re-executed on a different CPU.
In this scenario, when the sleeping thread finally completes removal of the
callout from some internal data structures, it may modify the wrong CPU's
data structures and thus leave them in an invalid state.

Obtained from: FreeBSD