libradius did not perform sufficient validation of received messages.
rad_get_attr(3) did not verify that the attribute length is valid before
subtracting the length of the Type and Length fields.  As a result, it
could return success while also providing a bogus length of SIZE_T_MAX -
2 for the Value field.

When processing attributes to find an optional authenticator,
is_valid_response() failed to verify that each attribute length is
non-zero and could thus enter an infinite loop.

Obtained from: FreeBSD