Fix an openssl vulnerability.
The BN_mod_sqrt() function, which computes a modular square root, contains
a bug that can cause it to loop forever for non-prime moduli. This function
is used when parsing certificates that contain certain forms of elliptic
curves.

A specially crafted certificate with invalid explicit curve parameters may
trigger an infinite loop, leading to a denial of service. Since certificate
parsing happens prior to verification of the certificate signature, any
process that parses an externally supplied certificate may be affected.