pam_krb5 authenticates the user by essentially running kinit(1) with the password, getting a `ticket-granting ticket' (tgt) from the Kerberos KDC (Key Distribution Center) over the network, as a way to verify the password.
Normally, the system running the pam_krb5 module will also have a keytab, a
key provisioned by the KDC. The pam_krb5 module will use the tgt to get a
service ticket and validate it against the keytab, ensuring the tgt is valid
and therefore, the password is valid.

However, if a keytab is not provisioned on the system, pam_krb5 has no way to
validate the response from the KDC, and essentially trusts the tgt provided
over the network as being valid.

Obtained from: FreeBSD